+++ src/main/java/insane96mcp/iguanatweaksreborn/data/generator/client/ISOBlockModelsProvider.java
@@ -19,6 +19,7 @@ public class ISOBlockModelsProvider extends BlockModelProvider {
         cross("solanum_neorossii", InsaneSO.location("block/solanum_neorossii")).renderType("cutout");
         flowerPotCross("potted_solanum_neorossii", InsaneSO.location("block/solanum_neorossii")).renderType("cutout");
         cube("fletching_table", ResourceLocation.parse("block/birch_planks"), ResourceLocation.parse("block/fletching_table_top"), ResourceLocation.parse("block/fletching_table_front"), ResourceLocation.parse("block/fletching_table_front"), ResourceLocation.parse("block/fletching_table_side"), ResourceLocation.parse("block/fletching_table_side"));
+        cubeColumn("burnt_log", InsaneSO.location("block/burnt_log"), InsaneSO.location("block/burnt_log_top"));
     }
 
     public BlockModelBuilder flowerPotCross(String name, ResourceLocation plant) {