+++ src/main/java/insane96mcp/iguanatweaksreborn/data/generator/client/ISOBlockStatesProvider.java
@@ -2,7 +2,9 @@ package insane96mcp.iguanatweaksreborn.data.generator.client;
 
 import insane96mcp.iguanatweaksreborn.module.farming.crops.Crops;
 import insane96mcp.iguanatweaksreborn.module.world.CyanFlower;
+import insane96mcp.iguanatweaksreborn.module.world.coalfire.CoalFire;
 import net.minecraft.data.PackOutput;
+import net.minecraft.world.level.block.RotatedPillarBlock;
 import net.minecraftforge.client.model.generators.BlockStateProvider;
 import net.minecraftforge.common.data.ExistingFileHelper;
 
@@ -15,5 +17,6 @@ public class ISOBlockStatesProvider extends BlockStateProvider {
     protected void registerStatesAndModels() {
         simpleBlock(CyanFlower.FLOWER.block().get());
         simpleBlock(Crops.SOLANUM_NEOROSSII.block().get());
+        logBlock((RotatedPillarBlock) CoalFire.BURNT_LOG.block().get());
     }
 }